mtest: Have all tests return non-0 on failure
Most tests are returning 0, even when they fail. This makes it difficult
to use a particular test in something like a git bisection which relies
on a non-zero return code to check correctness.

Potentially, this could also simplify the test scritps to just check the
return value instead of the correct output of " No Errors".